refactor(mm): M6-S2 extract staged-embed static helpers into neutral per-model modules
Second M6 step (non-destructive prep): the in-model wrappers imported the staged standalone-embed
classes purely to borrow static helpers, which would block deleting the staged embed models in
M6-S5. Extract those helpers into neutral per-model modules that both the in-model wrapper and the
(soon-to-be-deleted) staged embed class import -- single source of truth, no duplication, no cycle.

- mimo_v2_5/config_utils.py (new): normalize_vision_config + get_config_value, extracted from
  MiMoV2_5Embedding static methods. mimo_v2_5_inmodel.py + embedding.py both use it; the in-model
  wrapper drops its `from ...embedding import MiMoV2_5Embedding` import.
- qwen3_omni_moe/weights_mapping.py (new): create_visual_weight_mappings +
  create_audio_tower_weight_mappings, extracted from Qwen3OmniMoeThinkerEmbedding static methods
  (the staged-only text-embed mapping stays on the staged class -- the in-model wrapper reuses the
  AR embed_tokens). qwen3_omni_inmodel.py + thinker_embedding.py both use it; the in-model wrapper
  drops its `from ...qwen3_omni_thinker_embedding import Qwen3OmniMoeThinkerEmbedding` import.

After S1+S2 the in-model understanding path no longer imports any staged understanding module, so
M6-S5 can delete the staged path without touching in-model. Verified: all 6 modules import cleanly
on TPU pod (JAX_PLATFORMS=cpu).
